Update patch set 3
Patch Set 3:

(1 comment)

For all these datatypes that are not string, integer, boolean; an appropriate custom parser may be needed to ensure the values are correct. For the moment, I would use Instant and String as a last resort. But we will need to come in there and provide custom code to ensure those fields are correct. Otherwise, we will run into problems later very similar to the policy/engine where teams specify something legitimate but they fail to recognize that.

Patch-set: 3
Label: Code-Review=0