Update patch set 1
Patch Set 1:

I need to do the unit test for this review yet but I wanted to push it up for people to see. There is another review coming with a Java interface/API that the REST API code can work towards.

The reason for this is to keep the code that maintains database consistency behind the API. I'll explain at the stand up tomorrow.

Patch-set: 1