Update patch set 3
Patch Set 3:

I have some ideas on how we may get this to work:

I think we can safely say there are 2 levels of scaling: (1) at the kubernetes cluster level, (2) at the application level.

(2) cannot control anything from (1).    A necessary precondition for (2) is that independently a k8s cluster admin scales out the appropriate pdp pods (1).

A change in (1) does not propagate to (2), there are different abstractions.   But an operation in (1) have sideeffects that (2) can use for inferring the dynamic size of pooled resources.

Probably the minInstanceCount can be kept or perhaps renamed to desiredInstanceCount,
It is though probably think of it as an application level concept independent of kubernetes, although there is an indirect relationship due to the side effects I mentioned earlier.

kubernetes admin can and will change replica counts as they wish using their tooling, kubectl, helm, gui or others, as it is just a simple command operation that changes the replica count for a deployment.   This information does not flow directly to the application.   That's what I meant is difficult to keep in sync, any changes they introduce in the replica count with the application running in the containers.

But this will translate at runtime as the side effect I mentioned earlier with new PDPs registering (for scaling) or just stopping sending pdp-status notifications (de-scaling) so effectively you know what's the size of your pool of pdps at any given time.   The size of the pool can be very dynamic, and cannot be controlled by the application.   Based on that, PAP can decide how to allocate to groups and provide autoscalability within autoscalability, perhpas based on this minInstanceCount.   Point being is that the "min" in "minInstanceCount" is somewhat out of our application control.

Thoughts?

Patch-set: 3