Update patch set 7
Patch Set 7: Code-Review+1

Folks - in the interest of moving forward. Can we prioritize this? I think that we can open a task for any items to be cleaned up later. As long as there are no major showstoppers here. We need to get the API working for our clients.

Patch-set: 7