Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

We should follow TOSCA in the model even if what's coming in is not valid TOSCA. The model is designed to be as close to TOSCA as possible.

Also, I'm not sure what value hiding the fields from Swagger will have because the fields represent incoming fields (concept key -> name/version pair, and type-> policy type reference). Hiding the fields will not produce correct output on the swagger.

Patch-set: 1