Update patch set 1
Patch Set 1:

Chenfei, I see your point now,. we're talking about the value of the policy property and not its type definition. I think the correct way to model this is to have it as Map<PfCOnceptKey, Object>, where the PfConceptKey parent is the policy and the local name is the name of the property. This allows the property type to be checked. The type of the Object value can then be checked in validation

Patch-set: 1