Update patch set 1
Patch Set 1:

Thinking about it again, the correct mapping is Map<String, Object>, where the string is the policy name and the object is an object of the type in the policy definition. We can verify that.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0