Update patch set 2
Patch Set 2:

> (1 comment)

Pam, regarding this, I'd rather add policy-version in policyProvider than in dummyProvider as dummyProvider is just for tentative test. We'll change to use real modelProvider for DB. Adding policy-version in metadata is a general thing. So i think adding in policyProvider is more appropriate. Are you ok with that?

Patch-set: 2
Label: Code-Review=0