Update patch set 5
Patch Set 5:

> What bugs me is that all these dependencies that we could pick up
 > silently.   Looking at all the other classes definition in this
 > package, it seems to me that the class type of this field is very
 > inconsistent with the rest.   Would you agree?

With drools-pdp, for each DB table, we actually had two separate classes: one was a pure interface and the other was a JPA entity.  I wonder if we need the same thing here, only in this case we may actually want a third class, which is a pure POJO, with no JPA annotations.

Patch-set: 5