Update patch set 6
Patch Set 6: Code-Review-1

It just doesn't seem right to me the design decision of pulling the Pf* framework into the PAP-PDP network communication.

It is just this one single line of code that bothers me quite a bit:

     private List<PfConceptKey> supportedPolicyTypes;

as it comes with lots of baggage and introduces coupling with the *API* code on the PDP side.

I also see it very inconsistent with how all the other code in this "models-pdp" project is written - plain simple straightforward pojos. 

I don't think is maintanable to keep up with the changes in the pom and exclude them in the PDPs.  Let's say models start using eclipse EMF just to say, will have to remember to exclude.

Chenfei had issues with Pf* data leaking into serialization, so serialization/deserialization is tied to whatever goes into the PfConceptKey or super classes that should be of no concern on the PDP sides.

all of that, and at the end we just only need a very simple List<String>! (well perhaps a pair of 2 strings if we want to add the policy-type version to be exchanged)

Patch-set: 6
Label: Code-Review=-1