Update patch set 8
Patch Set 8:

Liam, I guess I don't understand where the duplication comes from and I agree I it is better to avoid duplication.  I question though if a network protocol should be a model, requestIds, statistics reports, healthchecks, statuses, .. are not models and an API user should never be allowed to do anything with those.    I guess I need to understand how the duplication claim comes about.  I just see them as 2 different animals, if we are mixing the networking protocols with the user models, may be we are not doing something right.

Patch-set: 8
Label: Code-Review=0