Update patch set 6
Patch Set 6:

According to tosca spec Sec. 3.8.1 and 3.9.1, I think we should add the dash since they are all taken as list. For policies, we already have them as list, but i don't know why policy_types and data_types have them as object, not list. The deserialization would be different with and without dash. With dash, it could be parsed as JsonArray. Without dash, it would be JsonObject. Maybe SDC missed that?

Patch-set: 6