Update patch set 4
Patch Set 4:

(1 comment)

Jim, on the BeanConfigurator, in my case I wanted the values in the output property fields to be null if they weren't in the map, not an empty string. That's why it wouldn't work for me.

Patch-set: 4
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0