Restructure for authorative models
Use authorative models instead of the JPA stored models
as the venacular models in the Policy Framework.

The next step is to pass the authorative objects over the
provider interface rather than the JPA obejcts. This change will
be made in the next review.

Minor changes to pdp objects to remove JPA things from them. These
object will be the authorative in-memory obejcts on the PDP side and
new JPA entity object will be introduced in a future review to persist
them.

Issue-ID: POLICY-1095
Change-Id: I40e7b713903980cb41bb315417111b67a9b17307
Signed-off-by: liamfallon <liam.fallon@est.tech>