Update patch set 2
Patch Set 2:

I wonder if it wouldn't be better to use interfaces instead of "authorative" models.  The interfaces would go in models-xxx, possibly along with a provider that provides implementations for JSON-decoding purposes.  The JPA provider can have its own implementations, that are returned from queries, and does the mapping from interface to JPA object when doing creates/updates.

Patch-set: 2