Refactor to authorative TOSCA serializtion
This review refactors the TOSCA support to use authorative
serialization and mapping.

It removes the JPA entities from the Proider interface

It brings the mapping support from and to authorative concepts out of
the JSON serialization classes directly into the JPA classes

It adapts the unit tests to work with the refactored structure.

Apologies for the review size but it all had to be done in a single block
of work.

Issue-ID: POLICY-1095
Change-Id: I4827d1dc67ef7aac98cba230ffcd79c6de71e805
Signed-off-by: liamfallon <liam.fallon@est.tech>