Update patch set 2
Patch Set 2:

Sorry to give you trouble Ajith, but I have reservations with regards to this .. I don't think this isn't properly models, it is for configuration parameters as apex uses it to get it from a json file, and not sure if it should be piggybacked to this code.   

I think this is currently used by apex-pdp only, not sure if xacml-pdp would use it for configuration.   In the drools-pdp, there's as of now no plan to use it, as we don't use json based configuration.

From the apex review, my comment was to used models-pdp structures to send status messages from PDP to PAP, which I think you are going to do next.   

For these parameters though used validation from json, my question is/was to see if you can extend or reuse some how in your parameters class the PdpStatus from models-pdp, so you can keep in sync, with the changes, as the fields the *Parameters seem to be duplicate of the PdpStatus ones.

I don't think keeping another copy of PdpStatus* alike of messages is maintainable in the long run.

Let me know your thoughts.

Patch-set: 2