Updates to models from scrum on 4/4
Moved ToscaPolicyIdentXxx classes from models-pdp to models-tosca,
and added methods to retrieve them from a ToscaPolicy.

Removed version and instance fields from PdpStatus.

Chose to leave "description" field in the PAP/PDP messages.  The
PDPs can ignore them or leave them null.  In a PdpUpdate message,
the description will be the description associated with the PdpGroup.
In a PdpStatus message, the PDP could choose to provide a description
of the PDP, itself (or the type of PDP).
Added comments to that effect.

Moved name, group, and subgroup fields into PdpMessage.

Fixed typos in comments.
Updated licenses.

Change-Id: Icd928f9a7630b838ad4b0b5556e899dc21b7872b
Issue-ID: POLICY-1542
Signed-off-by: Jim Hahn <jrh3@att.com>