Complete filters for Database Fetches
This review completes the implementaiton of the filters
for fetching policy types, policies, and PDP groups
from the database.

It also fixes bugs in Policy type creation.

Yaml in some of the policy type examples modified so that it is syntatically
correct.

Proeprties now stored as a blob in DB as they can be big.

Issue-ID: POLICY-1095
Change-Id: I6aef88ee2905afa58d778d82832f2b55d794fe9c
Signed-off-by: liamfallon <liam.fallon@est.tech>