Update patch set 1
Patch Set 1:

I don't see anything obvious for the tests to fail eyeballing the inspection.   Something I noticed is that the 2 errors found in the SoManagerTest is the error occurs first time we don't assert on "999"  (error).  In both cases, I think in those cases failing, it is expecting a valid response, so makes me suspicious of the http server querying against issue.   Just an observation.

Patch-set: 1