Update patch set 2
Patch Set 2: Code-Review+1

Yes, there's something equivalent in the pdp-d doing this, will change it to this method when I have a chance.   It could also be static or an utility provided that is accepting external group+subgroup.

As a sidenote, hopefully we don't keep the current implementation where we have to audit all policies in the PDPs every time there is a new policy being deployed and it is reworked in El Alto.  IRight now, the way I have it is that the policies that are not in the PDP-UPDATE list I undeploy them, which makes me a little bit nervous ..

I think this is going to be an inconvenience when doing lab testing either standalone or with integration and need manual procedures to continue testing, since to recover missing policies we will have to manually pre-fabricate the full list of policies.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1