Update patch set 2
Patch Set 2:

> Something isn't right here - actor.so and actor.vfc are already
 > there. So why are there new files?

This is because I didn't want to disturb the existing files which use named queries (incase we need to go back to named queries as a back up option). The new files for SO and VFC make use of custom queries. I can easily go back and delete the existing files (of SO and VFC) and replace them with these new ones, but then it will make it difficult to go back to named queries in case we need to. Let me know what to do.

Patch-set: 2