Update patch set 2
Patch Set 2:

First of all, NONE of the actors should be making AAI calls directly. The rules should automatically make the call to get the custom query so that all those details are available to all the actors. Now any policy can be written to specify an action on any VNF/VFModule for any actor/recipe. That was the ask - to clean up the AAI code.

Patch-set: 2