Update patch set 2
Patch Set 2:

> First of all, NONE of the actors should be making AAI calls
 > directly. The rules should automatically make the call to get the
 > custom query so that all those details are available to all the
 > actors. Now any policy can be written to specify an action on any
 > VNF/VFModule for any actor/recipe. That was the ask - to clean up


 > the AAI code.

None of the actors in the new code made a call to A&AI. Infact, in the new code, calls to A&AI are made from only a single function -- in ControlLoopEventManager -- that's it. Its a single line of code from where calls are made to A&AI. 
These SO and VFC classes need to change because the format of custom query response is very different now, and so to populate those requests, you have to change that part of the code, there is no way to get around that.

Patch-set: 2