Update patch set 2
Patch Set 2:

> First of all, NONE of the actors should be making AAI calls
 > directly. The rules should automatically make the call to get the
 > custom query so that all those details are available to all the
 > actors. Now any policy can be written to specify an action on any
 > VNF/VFModule for any actor/recipe. That was the ask - to clean up
 > the AAI code.

It is not possible to adapt to the new custom query by making changes only to the A&AI code. Given that the format of Custom Query Response is now entirely different, in order to populate the SO request and VFC Request, based on the new response, changes to the createRequest function  of SoServiceProvider is inevitable. 
In the new design, the CustomQuery response has all the information necessary for calculating targets in all use cases (VFC, SO, etc.). So in the new design, the calls are made only via the ControlLoopEventManager's getCqResponse() method. The ControlLoopOperation manager calls this method to get the CqResponse, and call the SO and VFC createRequest methods accordingly, by passing the custom query response.

Patch-set: 2
Reviewer: Gerrit User 5305 <5305@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0