Update patch set 1
Patch Set 1:

The problem is that the original "merge" of  policy-models-maven-stage-master/46/ (note the '46' instead of '48') that was before triggered a partial release, this is just trying to see if there's a work around for it, since it released some artifacts and others didn't.   It is a problem with the self-release tooling.  Not even sure if the tooling will do the right thing, if I bump the version again.   The problem is explained here:

https://jira.linuxfoundation.org/servicedesk/customer/portal/2/IT-17075

Note that as I mentioned the previous did release some subset of artifacts in 2.1.1 and others didnt' and remained in 2.1.0 even if the logs indicated that all of them were built.

What can yo do to help with that issue I reported?   

Can you remove the partial release of 2.1.1 of policy/models from nexus?    

Otherwise, it would be painful to propagate this new bump all over the policy repos again, and test it.   

What would guarantee that the next "merge" of the self-release tooling even if it's bump will work correctly and produce a full release instead of partial?

If you can please advice.

Patch-set: 1
Label: Code-Review=0