Update patch set 2
Patch Set 2: Code-Review-1

Thanks for the work.   The issue is that there is no junits, and we need quite a bit of Junits here to test this new functionality  (if there are problems in this area by the nature of the changes, they would manifest in obscure ways across several components I think).   

For policy/models coverage is currently close to 90% ..

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1