Update patch set 2
Patch Set 2:

(6 comments)

The reason for these copyTo methods was to allow seamless reading from and writing to the database. The way the DAO/JPA works (or worked a few years ago when I worked on this) means that it does not actually read everything from the database, it does a cached read where it only copies the top object of a hierarchy. It's only if you descend the object hierarchy that additional reads are done to read contained objects. The idea of the copyTo() method was to force a traversal of the entire object hierarchy tree and force a full copy of all the objects.

This code was originally designed to work towards Postgres and MariaDB may work differently.

We need to be sure that his behaviour is preserved.

Patch-set: 2
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0