Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)
This is a sample response which A&AI provided to me which is most complete. I have no control over what to put in this file. This file is primarily used to make sure that the structure of the JSON Custom Query response provided by A&AI is correct.

Patch-set: 4