Fix derivedFrom field
TOSCA does not have a version on derivedFrom fields on policies and
policy types. THerefore, we use the default value of 0.0.0 for versions
on these fields in the database.

When reading back out derivedFrom fields, we need to just take the name
of the field and not the ID of the field.

Issue-ID: POLICY-2075
Change-Id: I55e9c479047d6c1a9376a1713f4ddb2536476b29
Signed-off-by: liamfallon <liam.fallon@est.tech>