Update patch set 1
Patch Set 1: Code-Review+1

Just wondering, if derivedFrom now defaults to 0.0.0, will PDPs have problems doing a look-up of the type from which the type of interest is derived?  Wondering if derivedFrom should default to 1.0.0 instead of 0.0.0?  (Of course, if the derived-from version is something else, that won't work either.)
Maybe the code is right to begin with (until they add derivedFromVersion), and we should handle it in the invoking code?

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1