Update patch set 7
Patch Set 7:

> (1 comment)

I think that's much better, the fact that cleanInstance() has package visibility, does not prevent development in the future to do the wrong thing with it (since this is a library), and will confuse people looking at the code like me, since the combination of the 2 unsynchronized  methods mentioned, suggests potential race conditions.

As a note for the future, if cleanInstance() or something similar is only meant for testing, it would be better to move similar functionality under test, in future development, as there are some examples in the code where the singleton class is extended under src/test instead.

What I found most confusing since the constructor is parameterized, there would be parallel invocations such as 

Thread-1: CdsProcessorGrpcClient.getInstance(listenerA, propsA)
Thread-2: CdsProcessorGrpcClient.getInstance(listenerB, propsB) 

where thread-2 will get an instance that is using listenerA and propsA instead of listenerB, and propsB, as passed in the parameters.   It seems confusing to me, as my listenerB in thread-2 will never be invoked (instead it will be the listenerA) and the properties used would not be the one passed in.   The fact of having these parameters suggest to me that this should not be a singleton and perhaps a factory, which it seems that's what you're going to be changing.

Patch-set: 7