Update patch set 1
Patch Set 1:

Jim is this a timing issue with the new dmaap simulator JUnit test code? Pramod - this initially should be for the elalto branch anyway. Please cherry pick it to that branch so we can get an image to Marco. Jim's dmaap simulator changes are not on that branch.

Patch-set: 1