Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-1
> 
> until comments are addressed (either feedback response or code changes)

I uploaded additional changes this morning. I made changes in accordance with Jim's comment. I removed some of the redundant variable booleans that were in the code, not all could be removed due to the need to execute multiple checks to place messages in the callback related to the actor/target/recipe. Drools-Applications still builds correctly as well.

Patch-set: 2