Fix sonar issue affecting in drools-applications
Drools-applications has a serializable class that contains AaiCqResponse.
Sonar complained that the latter class must also be serializable, so
made the change in policy-models to fix that.  In addition, the latter
class contained a list of Object.  To ensure that those are also
serializable, the list was changed to a list of Serializable instead.

The latter change can be reverted, if it causes compilation issues
elsewhere.  However, it seems safer to enforce serializability in the
list rather than risk having a later change inject a non-serializable
object into the list.

Change-Id: Ia741a1a96ec4efcc63451e9af31c19b4fe99dd7d
Issue-ID: POLICY-2225
Signed-off-by: Jim Hahn <jrh3@att.com>