Update patch set 1
Patch Set 1: Code-Review+1

I've noticed that a lot of our logging is unnecessary, but Jorge is right in that this is quite important for debugging control loops. It would be good to sweep through all the API's including A&AI and cleanup the logging. Make sure we are dumping only what we need.

I'm ok with this coming in as long as we address it in Dec/Jan when we really cleanup actor code.

Patch-set: 1
Label: Code-Review=+1