Update patch set 3
Patch Set 3:

Maybe we can control this using logback.xml files. I have added one to the review here. Normally, the level of logging in the production code is set to "info", and then more verbose debugging can be turned on my changing the logging level in the logback.xml file on the machine where the system is being debugged.

Here, I have set the logging level to debug, which changes the level of logging back to the same as it was before this patch was introduced. However, now if it's too much for someone and they want to dial it back, they can set the logging level to info.

Patch-set: 3