Update patch set 7
Patch Set 7:

(1 comment)

Is the use of "Worker" in the class names too apex-pdp specific or can we use this concept in other PDPs as well? I guess so.

a "Worker" is a single thread of execution in a PDP. In a single PDP instance there can be many workers.

Patch-set: 7
Label: Code-Review=0