Actor prototype round #2
This only contains one revised actor, SDNC.  This actor combines code
from actor.sdnc, sdnc, and drools-applications.
Added operation Builder and simplified generation of futures.  Modified
the code to pass exceptions back via the future.
Allow executor service to be overridden.
Added ClosedLoopEventContext and moved the event and A&AI data inside,
eliminating the need for ControlLoopAaiData.
Modified code to throw exceptions when an actor or operation is not found,
or if the operation manager has not been started.  This makes it easier to
chain the calls together, using a single try/catch.
Moved classes around into different packages.
Modified to allow chaining of operations onto the end of an existing
future.

It will not compile as of yet.

Issue-ID: POLICY-1625
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I5beccd7c24cd4ee67b7a29b7d346afcd29c17cfc