Update patch set 4
Patch Set 4:

This is a general comment and is something to consider for the future but ti would be nice to do it in Frankfurt.

Maybe at some point we should consider the module structure and package naming structure. I think models-interactions was a catch-all we brought in when we lifted this code from drools-applications. One problem at the moment is that it's very difficult to know which module code is in because the package names are not mapped to the module structure. I always struggle with this.

For example the package "org.onap.policy.controlloop.actor.appc" is in model-interactions/model-actors/actor-appc

A suggestion would be to have a new module called "model-actor" and move the actor code into it. with a new package name hierarchy "org.onap.policy.models.actor........".

Also are actors exclusively for control loops, could they not be used in open loop operations. If that is the case, then we would not need "controlloop" in the package hierarchy.

Maybe we can move the code into new modules as it's being restructured and leave the old code in the existing location, and when all the new code is ready and not being used anymore, we can cut over to it.

Patch-set: 4