Implement validation and hierarchical get
NOTE: This review will require changes to the persistence.xml file in
other policy framework components.

This change brings in two changes:

1) Validation of the references betwen policies, policy types, and data
types

This change mans that additions to and changes to policies, policy
types, and data types must be structurally correct in the change coming
in and with whatever is already in the database

2) Hierarchical return of referenced entitites on policy types,
policies and data types.

When a policy, policy type, or data type is read from the database, the
entities referenced byt the returned policy, policy type, or data type
are returned as well.

Tests are amended for the stricter validation and structure.

Examples are corrected where they are incorrect.

Issue-ID: POLICY-1402
Change-Id: Ie6a4cb7ed336562338924079114df405f0ab889f
Signed-off-by: liamfallon <liam.fallon@est.tech>