Fixing issues with handling PolicyTypes having multiple versions
In ToscaServiceTemplate class, policyType/dataType is a Map<String, T>,
and this key String in the Map is taking in only the name, not version.
This can lead to issues when new versions are created.

In this fix, version is appended to the name and put as key. We could
also change the String to PfConceptKey so that name and version are
considered.

Change-Id: I0050c978ba6b29510fd4269c52641655c82b48f4
Issue-ID: POLICY-1402
Signed-off-by: a.sreekumar <ajith.sreekumar@est.tech>