More actor clean-up
Currently, Operator classes refer to Operation classes, and
vice versa, creating a dependency cycle.  In addition, there is
a slight problem in that if an operator is reconfigured, any
running operation may get inconsistent configuration data.
Modified the code to create Config objects that are passed to
the operation, instead of passing the Operator to the operations.
This solved both issues.

Replaceed makeOperator() with constructors.
Added parameter type to HttpActors.

Modified guard to get "ONAP" properties from its configuration,
as a default.

Changed setUp() to setUpBasic(), so "throws Exception" could be
removed, thus resolving a sonar issue.

Issue-ID: POLICY-1625
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I21eb8798acfbc636ff1bd8741b21c7278365b6e4